pyproject.toml
[project]
name = "cybergear"
version = "0.2.0"
description = "Python driver for the Xiaomi CyberGear brushless motor over CAN bus"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    "python-can>=4.4.2",
]

[project.optional-dependencies]
dashboard = [
    "textual>=0.80",
    "rich>=13.7",
]
dev = [
    "ruff>=0.6.2",
    "pytest>=8.0",
    "rich>=13.7.1",
    "mkdocs-material>=9.5",
]

[project.scripts]
cybergear-scan = "cybergear.cli.scan:main"
cybergear-dashboard = "cybergear.cli.dashboard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cybergear"]

[tool.ruff]
[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.ruff.lint]
extend-select = ["Q", "I"]
ignore = ["F821"]

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ty>=0.0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "hardware: requires a physical motor connected via CAN bus (deselect with '-m not hardware')",
]
